Move CodableFeedStore to its now file in production.
