+++ EssentialFeed/EssentialFeedTests/Cache Feed/CodableFeedStoreTests.swift
@@ -3,79 +3,6 @@ import EssentialFeed
 
 final class CodableFeedStoreTests: XCTestCase {
 
-    class CodableFeedStore: FeedStore {
-        private struct Cache: Codable {
-            let items: [CodableLocalFeedImage]
-            let timeStamp: Date
-
-            var localFeed: [LocalFeedImage] {
-                items.map{ $0.localFeedImage }
-            }
-        }
-
-        //DTO helper for internal implementation framework details
-        private struct CodableLocalFeedImage: Codable {
-            let id: UUID
-            let description: String?
-            let location: String?
-            let imageURL: URL
-
-            init(_ localFeedImage: LocalFeedImage) {
-                self.id = localFeedImage.id
-                self.description = localFeedImage.description
-                self.location = localFeedImage.location
-                self.imageURL = localFeedImage.imageURL
-            }
-
-            var localFeedImage: LocalFeedImage {
-                .init(id: id, description: description, location: location, imageURL: imageURL)
-            }
-        }
-
-        private let storeURL: URL
-
-        init(storeURL: URL) {
-            self.storeURL = storeURL
-        }
-
-        func insert(_ items: [LocalFeedImage], timeStamp: Date, completion: @escaping InsertionCompletion) {
-            do {
-                let encoder = JSONEncoder()
-                let cache = Cache(items: items.map(CodableLocalFeedImage.init), timeStamp: timeStamp)
-                let encodedData = try encoder.encode(cache)
-                try encodedData.write(to: storeURL)
-                completion(nil)
-            } catch {
-                completion(error)
-            }
-        }
-
-        func retrieve(completion: @escaping (RetrieveFeedResult) -> Void) {
-            guard let data = try? Data(contentsOf: storeURL) else {
-                return completion(.empty)
-            }
-            let decoder = JSONDecoder()
-            do {
-                let cache = try decoder.decode(Cache.self, from: data)
-                completion(.found(cache.localFeed, cache.timeStamp))
-            } catch {
-                completion(.failure(error))
-            }
-        }
-
-        func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-            guard FileManager.default.fileExists(atPath: storeURL.path) else {
-                return completion(nil)
-            }
-            do {
-                try FileManager.default.removeItem(at: storeURL)
-                completion(nil)
-            } catch {
-                completion(error)
-            }
-        }
-    }
-
     /// Clean any artefact Before/After running the test.
     private func setupEmptyState() {
         try? FileManager.default.removeItem(at: testSpecificStoreURL())
